% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{add_measure}
\alias{add_measure}
\title{Add the measurement variable to the query}
\usage{
add_measure(query, measure, rename = NULL)
}
\arguments{
\item{query}{An object created by start_query}

\item{measure}{The name of the measure}

\item{rename}{Optional. What should the measure variable returned in the data.frame be called}
}
\value{
A query object
}
\description{
Add the measurement variable to the query
}
\examples{
\dontrun{
pip_regs_age <- start_query("Personal Independence Payment", "PIP Registrations") |>
  add_measure("PIP Registrations") |>
  add_fields(c("Month", "Age (bands and single year)")) |>
  fetch()
 }

}
