% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{available_databases}
\alias{available_databases}
\title{View the available databases}
\usage{
available_databases(API_KEY)
}
\arguments{
\item{API_KEY}{Your statXplore account API key}
}
\value{
A dataframe of databases
}
\description{
View the available databases
}
\examples{
\dontrun{
available_databases("YOUR_API_KEY")
}

}
