% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{available_datasets}
\alias{available_datasets}
\title{Show the available datasets for a given database}
\usage{
available_datasets(database)
}
\arguments{
\item{database}{The name of a database returned by available_databases()}
}
\value{
A dataframe of datasets
}
\description{
Show the available datasets for a given database
}
\examples{

\dontrun{
available_datasets("Personal Independence Payment", "YOUR_API_KEY")
}

}
