% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_locat}
\alias{get_data_locat}
\title{Internal function to get the query name of the measure/field}
\usage{
get_data_locat(schema, field)
}
\arguments{
\item{schema}{The dataset scheme}

\item{field}{The field of interest}
}
\value{
A query name of the field
}
\description{
Internal function to get the query name of the measure/field
}
