% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_schema}
\alias{get_schema}
\title{Internal function to get the schema at an endpoint}
\usage{
get_schema(url, API_KEY)
}
\arguments{
\item{url}{The endpoint url}

\item{API_KEY}{Your account API key}
}
\value{
a dataframe of the schema
}
\description{
Internal function to get the schema at an endpoint
}
