% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the API Key}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{Your statXplore API Key, available from your statXplore account info}
}
\value{
No return, assigns the API to cache
}
\description{
Set the API Key
}
\examples{
set_api_key("YOUR_API_KEY")


}
