% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{start_query}
\alias{start_query}
\title{Start a statxplore query}
\usage{
start_query(database, dataset)
}
\arguments{
\item{database}{The name of the database}

\item{dataset}{The name of the dataset}
}
\value{
A query object
}
\description{
Start a statxplore query
}
\examples{
\dontrun{
pip_regs_age <- start_query("Personal Independence Payment", "PIP Registrations") |>
  add_measure("PIP Registrations") |>
  add_fields(c("Month", "Age (bands and single year)")) |>
  fetch()
 }

}
