% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{add_fields}
\alias{add_fields}
\title{Add the query fields}
\usage{
add_fields(query, fields, rename = NULL)
}
\arguments{
\item{query}{An object created by start_query}

\item{fields}{The names of the fields}

\item{rename}{Optional. What should the field variables returned in the data.frame be called}
}
\description{
Add the query fields
}
