% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{start_query}
\alias{start_query}
\title{Start a statxplore query}
\usage{
start_query(database, dataset, API_KEY)
}
\arguments{
\item{database}{The name of the database}

\item{dataset}{The name of the dataset}

\item{API_KEY}{Your statxplore account API key}
}
\description{
Start a statxplore query
}
